+++ vite.config.ts
@@ -1,5 +1,4 @@
 import { defineConfig, loadEnv } from 'vite';
-import vitePluginImp from 'vite-plugin-imp';
 import { viteMockServe } from 'vite-plugin-mock';
 import react from '@vitejs/plugin-react-swc';
 import { resolve } from 'path';
@@ -37,19 +36,7 @@ export default defineConfig(({ command, mode }) => {
       //   cert: fs.readFileSync('CertKeys/agent2-cert.pem'),
       // },
     },
-    plugins: [
-      react(),
-      // vitePluginImp({
-      //   optimize: true,
-      //   libList: [
-      //     {
-      //       libName: 'antd',
-      //       style: (name) => `antd/es/${name}/style`,
-      //     },
-      //   ],
-      // }),
-      viteMockServe({ mockPath: './mock' }),
-    ],
+    plugins: [react(), viteMockServe({ mockPath: './mock' })],
     css: {
       preprocessorOptions: {
         less: {